Add user authentication with login and registration
- Implement user registration with username/password validation
- Implement secure login with password hashing (werkzeug)
- Add SQLite database setup with users table
- Create CLI interface for authentication workflows
- Move project files to project_v1/ directory structure
- Add VS Code settings for Python path configuration
- Update .gitignore to exclude database files

Closes #1
