--- capstone_core/db.py
@@ -1,18 +1,0 @@
-# capstone_core/db.py
-import sqlite3
-
-DB_PATH = "data/capstone.db"
-
-def get_conn():
-    return sqlite3.connect(DB_PATH)
-
-def init_db():
-    with get_conn() as conn:
-        conn.execute("""
-        CREATE TABLE IF NOT EXISTS users (
-            id INTEGER PRIMARY KEY AUTOINCREMENT,
-            username TEXT NOT NULL UNIQUE,
-            password_hash TEXT NOT NULL,
-            created_at TEXT NOT NULL DEFAULT (datetime('now'))
-        );
-        """)
